% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_image.R
\name{mcr_autoconstrain}
\alias{mcr_autoconstrain}
\title{Automatic peak detection}
\usage{
mcr_autoconstrain(x)
}
\arguments{
\item{x}{an image loaded with \code{\link{mcr_load_image}}}
}
\value{
A \code{\link[tibble]{tibble}} containing the positions of peak maxima and half-height-width both to the left and to the right of the peak
}
\description{
This function automatically determines the position of peaks 
and their half-height-width in brightness density distribution of an image.
This is key input information necessary to assign brightness-niveaus to
certain (mineral) phases an can be used in \code{\link{mcr_herd_minerals}}
}
\examples{
myimage <- mcr_load_image(system.file("extdata", "testim.png", package = "mincountr"))
mcr_autoconstrain(myimage)
 
}
