% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_image.R
\name{mcr_herd_minerals}
\alias{mcr_herd_minerals}
\title{Calculate phase area share}
\usage{
mcr_herd_minerals(x, lhs, rhs)
}
\arguments{
\item{x}{an image loaded with \code{\link{mcr_load_image}}}

\item{lhs}{a vector with the left-hand-side position of peaks observed in \code{\link{mcr_inspect_phases}}}

\item{rhs}{a vector with the right-hand-side position of peaks observed in \code{\link{mcr_inspect_phases}}}
}
\value{
A \code{\link[tibble]{tibble}} containing the number of pixels and their relative share of total image area per phase
}
\description{
calculate the percentage share of mineral phases of the area of
an image.
}
\examples{
myimage <- mcr_load_image(system.file("extdata", "testim.png", package = "mincountr"))
mypeaks <- mcr_autoconstrain(myimage)
mcr_herd_minerals(myimage, mypeaks$x1, mypeaks$x2)

}
