% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_image.R
\name{mcr_inspect_assignment}
\alias{mcr_inspect_assignment}
\title{Plot group assignment of a phase in an image}
\usage{
mcr_inspect_assignment(x, lhs, rhs)
}
\arguments{
\item{x}{an image loaded with \code{\link{mcr_load_image}}}

\item{lhs}{a vector with the left-hand-side position of peaks observed in \code{\link{mcr_inspect_phases}}}

\item{rhs}{a vector with the right-hand-side position of peaks observed in \code{\link{mcr_inspect_phases}}}
}
\description{
Plots a false color image using groups of brightness that can 
either be determined from a graphical inspection of the result of
\code{\link{mcr_inspect_phases}} or automaticalle determined by
\code{\link{mcr_herd_minerals}}
}
\examples{
myimage <- mcr_load_image(system.file("extdata", "testim.png", package = "mincountr"))

# (Semi-)Automatic approach:
mypeaks <- mcr_autoconstrain(myimage)
mcr_inspect_assignment(myimage, mypeaks$x1, mypeaks$x2)

# Manual assignments of brightnessgrouplimits:
mcr_inspect_assignment(
myimage,
lhs = c(0, 0.3, 0.5, 0.92),
rhs = c(0.05, 0.45, 0.65, 1)
)

}
