% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_image.R
\name{mcr_inspect_phases}
\alias{mcr_inspect_phases}
\title{Plot an images brightness density distribution}
\usage{
mcr_inspect_phases(x)
}
\arguments{
\item{x}{an image loaded with \code{\link{mcr_load_image}}}
}
\description{
Inspecting the density distribution of an electron 
microscope image of minerals effectively allows to identify different phases
as their distinct brightness is a key giveaway of their different chemical
composition (and conductivity).
}
\examples{
myimage <- mcr_load_image(system.file("extdata", "testim.png", package = "mincountr"))
mcr_inspect_phases(myimage)

}
