% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_image.R
\name{mcr_load_image}
\alias{mcr_load_image}
\title{Load image from file or URL}
\usage{
mcr_load_image(file)
}
\arguments{
\item{file}{path to file of URL}
}
\value{
an object of class \code{\link[imager]{cimg}}
}
\description{
simple wrapper around \code{\link[imager]{load.image}} included
here for convenience and consitency reasons only.
}
\examples{
mcr_load_image(system.file("extdata", "testim.png", package = "mincountr"))

}
